WIP: Add support for per-monitor DPI awareness (6)

+++ Src/AboutDlg.cpp
@@ -59,9 +59,10 @@ public:
 	afx_msg BOOL OnEraseBkgnd(CDC* pDC);
 	afx_msg HBRUSH OnCtlColor(CDC* pDC, CWnd* pWnd, UINT nCtlColor);
 	afx_msg void OnBnClickedWWW(NMHDR *pNMHDR, LRESULT *pResult);
+	afx_msg LRESULT OnDpiChanged(WPARAM wParam, LPARAM lParam);
 
 private:
-	virtual void UpdateDpi(int dpi) override;
+	void RecreateResources();
 
 	CAboutDlg *const m_p;
 	ATL::CImage m_image;
@@ -76,6 +77,7 @@ BEGIN_MESSAGE_MAP(CAboutDlg::Impl, CTrDialog)
 	ON_WM_ERASEBKGND()
 	ON_WM_CTLCOLOR()
 	ON_NOTIFY(NM_CLICK, IDC_WWW, OnBnClickedWWW)
+	ON_MESSAGE(WM_DPICHANGED, OnDpiChanged)
 END_MESSAGE_MAP()
 
 CAboutDlg::Impl::Impl(CAboutDlg *p, CWnd* pParent /*= nullptr*/)
@@ -105,10 +107,8 @@ BOOL CAboutDlg::Impl::OnInitDialog()
 		// FIXME: LoadImageFromResource() seems to fail when running on Wine 5.0.
 	}
 
-	UpdateDpi(GetDpi());
-	
-	GetDlgItem(IDC_VERSION)->SetFont(&m_font);
-	GetDlgItem(IDC_GNU_ASCII)->SetFont(&m_font_gnu_ascii);
+	RecreateResources();
+
 	::SetDlgItemTextA(m_hWnd, IDC_GNU_ASCII, gnu_ascii);
 
 	String link;
@@ -166,25 +166,35 @@ void CAboutDlg::Impl::OnBnClickedWWW(NMHDR *pNMHDR, LRESULT *pResult)
 	ShellExecute(nullptr, _T("open"), pNMLink->item.szUrl, nullptr, nullptr, SW_SHOWNORMAL);
 }
 
-void CAboutDlg::Impl::UpdateDpi(int dpi)
+LRESULT CAboutDlg::Impl::OnDpiChanged(WPARAM wParam, LPARAM lParam)
 {
-	m_dpi = dpi;
+	CTrDialog::OnDpiChanged(wParam, lParam);
+	RecreateResources();
+	Invalidate();
+	return 0;
+}
 
-	auto pointToPixel = [dpi = GetDpi()](int point) { return MulDiv(point, dpi, 72); };
+void CAboutDlg::Impl::RecreateResources()
+{
+	if (!GetDlgItem(IDC_VERSION))
+		return;
+	
+	const int dpi = GetDpi();
 
-	LOGFONT lfv = { 0 };
-	lfv.lfHeight = -pointToPixel(10);
+	LOGFONT lfv{};
+	DpiAware::GetPointLogFont(lfv, 10, _T("Tahoma"), dpi);
 	lfv.lfWeight = FW_NORMAL;
-	_tcscpy_s(lfv.lfFaceName, _T("Tahoma"));
 	m_font.DeleteObject();
 	m_font.CreateFontIndirect(&lfv);
 
-	LOGFONT lf = { 0 };
-	lf.lfHeight = -pointToPixel(14);
+	LOGFONT lf{};
+	DpiAware::GetPointLogFont(lf, 14, _T("Courier New"), dpi);
 	lf.lfWeight = FW_BOLD;
-	_tcscpy_s(lf.lfFaceName, _T("Courier New"));
 	m_font_gnu_ascii.DeleteObject();
 	m_font_gnu_ascii.CreateFontIndirect(&lf);
+
+	GetDlgItem(IDC_VERSION)->SetFont(&m_font);
+	GetDlgItem(IDC_GNU_ASCII)->SetFont(&m_font_gnu_ascii);
 }
 
 CAboutDlg::CAboutDlg() : m_pimpl(new CAboutDlg::Impl(this)) {}
